DMA Queue stuff
Three things really: trying to make the DMAs RAM safe (not yet), adding a crash when the game does 128kb DMA when we turned that off, and adding a fix to the registers during the 128kb fixing process if we have that on. All of those are optional
I did try to put a lot of effort into the debug code. Hopefully it'll suffice