size optimization to GetFloorPosition
The original code uses a LUT to find the chunk we're on, but the same calculation can be done with a left shift by 7. This saves a lot of space, though is ever so slightly slower by 2 cycles.
I might take a look into FindFloor in general soon. I think improving that could allow some more complex object code, or at the very least improve the scattered rings performance