fixed bug with KosM queue
The code handles moving the queue in two longwords, when the data is a long and word. This can be fine if the amount of queues is an even number... but it isn't. That makes the queue clearing unsafe, though it's only noticeable in higher queue numbers.
I will give credit where it's due though, it's more optimal then the S3K version. That one was far too safe, and would be way too slow if the KosM was any bigger then 4.

I also decreased the amount of queues from 32 to 20. You can change that back if the game needs more. That's kind of the whole point of that label, actually.