optimized movesprite since we're using d1
Made a macro to make MoveSprite easier to inline, alongside genuine code speedup. The MoveSprite it's based on was pretty similar to what vlad made, but removed two ext instructions thanks to movem.w
The macro itself is an overengineered mess, particularly with how it handles falling velocity
Implementation of the macro should be simple enough, though. Since the macro flags assume they're 0 if they aren't set, you can have it with no flags and it'll work (gives you SpeedToPos)