Update CalcSine.asm
Older versions of SCE featured an optimized version of CalcSine, which got reverted likely due to being created by a now controversial person, and generally not being a massive game changing speedup.
Due to a bit of guilt I have around that code, and a new philosophy I've developed for situations like this, I decided to rewrite and outclass that original code, effectively obsoleting it.
My version is equal in speed to that version if you use it normally, but allows you to dance with unsafe register use, if your code clears the high byte of d1.
I also included a WIP macro version, which uses an aX register of your choice (I tested with a1). That versions code is slower, but is easy to inline which overall saves cycles.