slightly improved loading times for title card
Loading the HUD graphics so early got in the way of loading the title card graphics, and LoadLevelLoadBlock could've loaded while the title card was still animating. This change fixes both, saving a few frames of loading time
The code also now changes the players transparency palette, to mark what type of loading the level init is doing. This serves as a small debug thing for emulators, to roughly get an idea of how long it takes to load the level graphics