Speedups and fixes to level collision
Optimizing floor collision is a pretty crucial thing. You could save a decent amount of space by making the macro copies there own routine, like GetFloorPosition
the change in level data is for testing loop collision. There's a particular bug with the test, but I confirmed the code changes had nothing to do with that; Probably an issue with the level collision itself