organizing RAM and change to player DPLCs
Players now use a new variable I added a while ago, instead of hardcoded areas in RAM. This saves RAM and makes adding more characters easier, but requires the a1 register (Tails' tails could get away with it, but just to be sure). This could create minor issues for legacy code, but it's not a big deal. The original labels are defined at the end of Object RAM, for backwards compatibility
I also included a new macro for making RAM evenly spaced. It's a basic modification of evensimple (which had to be fixed due to having glaring issues)
might make another folder for backcompat, since it's getting pretty cluttered