optimization, documentation, code flexibility
the main change is how the Max_speed and co's variables work. A new one was added (Jump_height), and the hardcoded speed changes have been made into a routine. I sighted a tutorial as credit, but in truth I wrote most of this myself, mainly using it as a reference for the lea instruction because I forgot how pc relatives work (funny considering calcsine depends on it haha...), though it's the same idea and it's what I was thinking of when making it, so, I'd say it counts.
Jump_height was added to make different jumping heights easier to implement (I'm considering making Sonic_Jump more of a Player_Jump if you will), and because the new routine uses bit shift multiplication meaning it was limited to x8 per variant of speed, so it makes sense to use that extra space.
Thinking of merging extra variables like this into Object RAM, into objects like Tails' tails. That could overall save RAM, but seems a little hard to work with, especially for legacy labels
Also optimized some CalcSine uses, and added .angd1, which assumes angle is already in d1 with the high byte cleared. That saves like 4 cycles, only do it if you're being stupid in the middle of the night
It is completely possible I broke something somewhere. Tails in particular had some tragic things happen due to some find and replace errors