BCD lives counter
On my quest to speed up Vint, I tested bcd instructions with the life counter. This speeds up HUD_Lives since it doesn't have to do a binary conversion, since it's, y'know, already converted. Though depending on the implementation, it could overall be a net loss. Ah well, that's what commit reverting is for. I plan to do the same with the score, once I become more comfortable with implementing that. It'll be much faster, and also easier to read in a RAM editor. Might even save a byte or two.
Also changed some methods of gaining 1ups, so they branch to a routine called AddExtraLife, and game overs now happen when you die with 0 lives instead of 1, similar to the likes of Sonic Advance