6 button controller state update
Honestly surprised it didn't click to me until now. Instead of leaving CtrlXState as is, it check what it was the previous frame and does the appropriate handler. This gives CtrlXState a more active use for regular controllers, although it is an unusual edgecase that doesn't effect real hardwa- ...oh. Uhh, apparently it can effect broken controllers. Huh.
Keep in mind, this version can run into an error if you were holding up and down when on 6pad, when switch to a 3pad on the same frame while still holding those buttons down. However, I think that scenario is kind of retarded, and even if it were a massive deal, the regular controllers are just too simple to fix it...
I kept the old version(s) commented. An error is an error, and if you don't like that, I encourage you to find a method that best suits you. That, and having pseudo-controller modes to fix an obscure error is kind of funny