hardware detection
This update is certainly something. There's a lot of things left to implement that this introduces, and I have to set up a proper way to test it all. On the bright side, it gives me stuff to do
During the process of cleaning this all up, I'll be working on engine documents, and comparisons of features to real hardware and emulation. That should help make developers better understand the more exotic features I've introduced