MegaSD detection
MD+ here we come! Ironically, MD+ is a lot closer to an MSU system then MSU-MD, and since it's separate from the CD hardware, it allows more Mode 1 stuff
I also have labels for how the SD handles the Sega/SSF2 Mapper. Sadly that might not be of any use...