constants
I've ran into an issue. AS needs constants and macros to be defined as early as possible to allow all code to use them. However, constants especially is gonna get really bloated from that, and disorganized. So, I'm trying to make constants easier to organize and keep up the top. Macros shouldn't be too annoying