Changes to start of header and vector data
Enabled External interrupts for testing purposes. It's intended to play a sound for now
changed the SRAM header. CartRAM_Type has a bitfield which might be useful outside of SRAM use, so I wanted to make it easier to access. I also noticed that one of the bits is only set when SRAM is active (now labelled as E), so better documentation